moved location_codes to utilites, cosmetic changes
