+++ src/tfr_navigation/src/navigation_action_client.cpp
@@ -1,7 +1,7 @@
 #include<ros/ros.h>
 #include<navigation_client.h>
 #include<tfr_msgs/NavigationAction.h>
-#include<tfr_msgs/location_codes.h>
+#include<tfr_utilities/location_codes.h>
 
 /**
  *  Main entry point for test class of navigation action server, eventually
@@ -14,21 +14,9 @@ int main(int argc, char** argv)
     ros::NodeHandle n;
 
     NavigationClient client("navigation_action_server");
-    //test preemption
-    client.navigate(tfr_msgs::LocationCode::MINING);
-//    client.stop_all();
-    
-    //test happy path mining
-//    client.navigate(tfr_msgs::LocationCode::DUMPING);
-    for (int i = 0; i < 16; i++)
-    {
-        ros::spinOnce();
-        ros::Duration(0.25).sleep();
-    }
-
+    client.navigate(tfr_utilities::LocationCode::MINING);
     
     //test happy path dumping
-//    client.navigate(tfr_msgs::LocationCode::DUMPING);
     ros::spin();
     return 0;
 }
