+++ src/tfr_navigation/src/navigation_action_server.cpp
@@ -22,4 +22,3 @@ int main(int argc, char** argv)
     ros::spin();
     return 0;
 }
-
